refactor Structure Tab merge_rtss
